Verklapper en hr in prosemirror
