+++ assets/js/forum/editor.ts
@@ -1,60 +1,23 @@
-import {MenuItem} from "prosemirror-menu"
-import {EditorState, NodeSelection} from "prosemirror-state"
+import {EditorState} from "prosemirror-state"
 import {EditorView} from "prosemirror-view"
 import {Node, Schema} from "prosemirror-model"
 import {schema} from "prosemirror-schema-basic"
 import {addListNodes} from "prosemirror-schema-list"
 import {buildMenuItems, exampleSetup} from "prosemirror-example-setup"
-import {openPrompt, TextField} from "prosemirror-example-setup/src/prompt"
 import ctx from "../ctx";
-import {bbBlockSpec, blocks} from "./schema";
+import {addBbBlock, buildBbBlockMenu} from "./bbBlockSchema";
+import {addBbVerklapper, buildBbVerklapperMenu} from "./bbVerklapperSchema";
 
 // Mix the nodes from prosemirror-schema-list into the basic schema to
 // create a schema with list support.
 const mySchema = new Schema({
-	nodes: addListNodes(schema.spec.nodes as any, "paragraph block*", "block")
-		.addBefore("image", "bb-block", bbBlockSpec),
+	nodes: addBbVerklapper(addBbBlock(addListNodes(schema.spec.nodes as any, "paragraph block*", "block"))),
 	marks: schema.spec.marks
 })
 
-const blockType = mySchema.nodes["bb-block"]
-
-const canInsertBlock = state => {
-	const {$from} = state.selection
-
-	for (let d = $from.depth; d >= 0; d--) {
-		const index = $from.index(d)
-		if ($from.node(d).canReplaceWith(index, index, blockType)) return true
-	}
-
-	return false;
-}
-
 const menu = buildMenuItems(mySchema)
-
-Object.entries(blocks).forEach(([type, fields]) => menu.insertMenu.content.push(new MenuItem({
-	title: "Insert " + type,
-	label: type.charAt(0).toUpperCase() + type.slice(1),
-	enable: canInsertBlock,
-	run(state, _, view) {
-		let attrs = null
-
-		if (state.selection instanceof NodeSelection && state.selection.node.type == blockType) {
-			attrs = state.selection.node.attrs
-		}
-
-		openPrompt({
-			title: attrs && attrs.id ? "Update: " + attrs.type : "Invoegen: " + type,
-			fields: Object.fromEntries(
-				fields.map(field =>
-					[field, new TextField({label: field, required: true, value: attrs && attrs[field]})])),
-			callback(attrs) {
-				view.dispatch(view.state.tr.replaceSelectionWith(blockType.createAndFill({type, ...attrs})))
-				view.focus()
-			}
-		})
-	}
-})))
+buildBbBlockMenu(menu)
+buildBbVerklapperMenu(mySchema, menu)
 
 ctx.addHandler('.pm-editor', el => {
 	const input = document.querySelector<HTMLInputElement>('#' + el.dataset.prosemirrorDoc);
@@ -76,6 +39,3 @@ ctx.addHandler('.pm-editor', el => {
 		}
 	})
 })
-
-
-
