+++ assets/scss/module/_formulier.scss
@@ -46,6 +46,27 @@
 	padding: 2em;
 }
 
+.ProseMirror .pm-verklapper {
+	border: 1px solid rgba(0,0,0,.125);
+	border-radius: 0.25rem;
+	padding: 1.25rem;
+
+	&::before {
+		margin: -1.25rem -1.25rem 0 -1.25rem;
+		content: 'Verklapper';
+		display: block;
+		padding: .25rem .5rem;
+		font-size: .875rem;
+		line-height: 1.5;
+		border-radius: .2rem;
+		color: #fff;
+		background-color: #6c757d;
+		border-color: #6c757d;
+		text-align: center;
+		width: calc(100% + 1.25rem);
+	}
+}
+
 input[type="file"] {
   background-color: transparent;
   border: none;
