+++ package.json
@@ -24,6 +24,7 @@
 		"@types/mini-css-extract-plugin": "^0.2.0",
 		"@types/node": "^10.12.19",
 		"@types/optimize-css-assets-webpack-plugin": "^1.3.3",
+		"@types/prosemirror-menu": "^1.0.3",
 		"@types/prosemirror-model": "^1.11.2",
 		"@types/prosemirror-schema-basic": "^1.0.2",
 		"@types/prosemirror-schema-list": "^1.0.2",
